+++ Fable.Helpers.Mqtt.fs
@@ -2,8 +2,32 @@
     
 open Fable.Core
 open Fable.Import.JS
+open Fable.Import.Mqtt
 
-[<Import("*","mqtt")>]
-let mqtt : Fable.Import.mqtt.Globals = failwith "JS only"
+type [<KeyValueList>] ClientOptions =
+    | Keepalive of float
+    | ClientId of string
+    | ProtocolId of string
+    | ProtocolVersion of float
+    | Clean of bool option
+    | ReconnectPeriod of float
+    | ConnectTimeout of float
+    | Username of string
+    | Password of string
+    | IncomingStore of Store
+    | OutgoingStore of Store
+    | Will of obj
+    interface IClientOptions
+
+type [<KeyValueList>] SecureClientOptions =
+    | KeyPath of string
+    | CertPath of string
+    | RejectUnauthorized of bool
+    interface IClientOptions
 
 let callback (handler: string -> 'a -> unit) = System.Func<string,string,unit>(fun topic msg -> JSON.parse msg :?> 'a |> handler topic)
+
+[<Import("*","mqtt")>]
+let mqtt : Static = jsNative
+
+let inline connect url opts = mqtt.connect (url, unbox opts)